Adding sha1_io.c/h: A wrapper interface for performing posix file I/O with SHA1 as a CRC mechanism.
When writing file formats using SHA1 as CRC, its is handy to
have SHA1_Update() to be applied to every write(). so that an
SHA1 hash can be calculated for that data and used as an CRC check.

Therefor this interface is created to wrap the code used to do this.