lockfile: Only add a lock once to the linked list.
remove_from_list() is still there, but should not always be used.
Most of the time, delete from the linked list is unnecessary.
If a struct lockfile's storage is valid throughout the entier life of a program.
And is used (for example) to lock 3 different files in sequence or something.
removing the lockfile from the linked list each time a call to 'commit'
or 'release' is made, and then inserting again on each hold is unnecessary.
Most of the time a lockfile has a longer lifetime than a function's stack frame.

To implement this, a variables is added to the struct to keep track of if
it exists in the list or not. And when calling hold_lock(): insert it if it's not added.

If hover, a lockfile is created in a local function and the caller wishes to
forget about it when leaving the local scope, a separated function with the purpose
of removing the lockfile from the linked list can be implemented, and the caller
should be responsible of calling that function in that case.