lockfile.c: Do not remove O_EXCL when trying to force taking a lock in open_lock()
By making open(2) not fail if the file exists, a race condition can happen.
If 2 processes finds out at the same time that the file exists AND the
file is older than MAX_LOCK_TIME
Both processes will open the file without O_EXCL and both will therefore succeed.

To fix this, use unlink(2) to remove the file first. then a call to open(2) with
O_EXCL flag, open(2) is an atomic operation, and only 1 process will be able to
create the file again and the other process will fail.